Added FileChecker class
